+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Full documentation for rocSOLVER is available at the [rocSOLVER documentation](h
     - LACGV_64
     - LARF_64
     - LARFG_64
+    - GEQR2_64 (with batched and strided\_batched versions)
+    - GEQRF_64 (with batched and strided\_batched versions)
 
 ### Optimized
 ### Changed